requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Django==4.2.7
django-environ==0.10.0
django-filter==23.3
djangorestframework==3.14.0
drf-spectacular==0.26.5
inflection==0.5.1
jdatetime==4.1.1
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
Markdown==3.5.1
psycopg2-binary==2.9.9
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
rpds-py==0.12.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
